# apps/bilgeapi/docs/openapi/bilgeapi_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BilgeAPI",
    "description": "Independent Incident Intake, Diagnostic and Repair-Orchestration API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/v1/repair-requests": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Create Repair Request",
        "description": "Creates a pending repair request for a diagnostic run.",
        "operationId": "create_repair_request_v1_repair_requests_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The associated diagnostic ID",
              "title": "Diagnostic Id"
            },
            "description": "The associated diagnostic ID"
          },
          {
            "name": "requested_by",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requested By"
            }
          },
          {
            "name": "risk_score",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Risk Score"
            }
          },
          {
            "name": "risk_reason",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Risk Reason"
            }
          },
          {
            "name": "approval_required",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Approval Required"
            }
          },
          {
            "name": "approval_status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ApprovalStatus",
              "default": "PENDING"
            }
          },
          {
            "name": "approved_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approved By"
            }
          },
          {
            "name": "approved_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approved At"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Repairs"
        ],
        "summary": "List Repair Requests",
        "description": "Lists all repair requests.",
        "operationId": "list_repair_requests_v1_repair_requests_get",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepairRequestResponse"
                  },
                  "title": "Response List Repair Requests V1 Repair Requests Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repair-requests/json": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Create Repair Request Json",
        "description": "Creates a pending repair request using JSON body.",
        "operationId": "create_repair_request_json_v1_repair_requests_json_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/diagnostics/{diagnostic_id}/repair-requests": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Create Repair Request From Diagnostic",
        "description": "Creates a repair request by evaluating the diagnostic run results and calculating deterministic risk.",
        "operationId": "create_repair_request_from_diagnostic_v1_diagnostics__diagnostic_id__repair_requests_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          },
          {
            "name": "requested_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional custom requested_by, defaults to authenticated user ID",
              "title": "Requested By"
            },
            "description": "Optional custom requested_by, defaults to authenticated user ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repair-requests/{id}": {
      "get": {
        "tags": [
          "Repairs"
        ],
        "summary": "Get Repair Request",
        "description": "Retrieves a single repair request by ID.",
        "operationId": "get_repair_request_v1_repair_requests__id__get",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repair-requests/{id}/approve": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Approve Repair Request",
        "description": "Approves a repair request. Does not dispatch webhook.",
        "operationId": "approve_repair_request_v1_repair_requests__id__approve_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repair-requests/{id}/reject": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Reject Repair Request",
        "description": "Rejects a repair request with a reason.",
        "operationId": "reject_repair_request_v1_repair_requests__id__reject_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairRejectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repair-requests/{id}/dispatch": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "summary": "Dispatch Repair Request",
        "description": "Dispatches the approved (or auto-approved) repair request to the target webhook.",
        "operationId": "dispatch_repair_request_v1_repair_requests__id__dispatch_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RepairDispatchRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Dispatch Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/webhooks/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Test Webhook",
        "description": "Sends a test payload to the target webhook URL using the full SSRF, signature, and timeout flow.",
        "operationId": "test_webhook_v1_webhooks_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED"
        ],
        "title": "ApprovalStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepairDispatchRequest": {
        "properties": {
          "adapter": {
            "type": "string",
            "title": "Adapter",
            "default": "webhook"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "title": "RepairDispatchRequest"
      },
      "RepairRejectionRequest": {
        "properties": {
          "rejection_reason": {
            "type": "string",
            "title": "Rejection Reason"
          }
        },
        "type": "object",
        "required": [
          "rejection_reason"
        ],
        "title": "RepairRejectionRequest"
      },
      "RepairRequestCreate": {
        "properties": {
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_reason": {
            "type": "string",
            "title": "Risk Reason"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required",
            "default": true
          },
          "approval_status": {
            "$ref": "#/components/schemas/ApprovalStatus",
            "default": "PENDING"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          }
        },
        "type": "object",
        "required": [
          "requested_by",
          "risk_score",
          "risk_reason"
        ],
        "title": "RepairRequestCreate"
      },
...
    },
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}